convert Roche model numbers into model names for Bluetooth driver
