+++ lib/drivers/bluetoothLE/bluetoothLEDriver.js
@@ -25,6 +25,7 @@ import structJs from '../../struct';
 import common from '../../commonFunctions';
 import uploadDataPeriod from '../../../app/utils/uploadDataPeriod';
 import api from '../../core/api';
+import getModelName from '../roche/models';
 
 const struct = structJs();
 
@@ -113,7 +114,7 @@ module.exports = (config) => {
        if(cfg.deviceInfo.name.startsWith('meter+')) {
          cfg.deviceInfo.serial = cfg.deviceInfo.model + cfg.deviceInfo.name.slice(6);
          cfg.deviceInfo.deviceId += `-${cfg.deviceInfo.serial}`;
-         data.deviceModel = cfg.deviceInfo.model;
+         data.deviceModel = getModelName(cfg.deviceInfo.model);
        } else if(env.electron){
          cfg.deviceInfo.deviceId += `-${remote.getGlobal('bluetoothDeviceId')}`;
        } else {
@@ -401,6 +402,12 @@ module.exports = (config) => {
        tzName: cfg.timezone,
        version: cfg.version,
      };
+
+     if(cfg.deviceInfo.name.startsWith('meter+')) {
+       // For Roche meters, we want to store the model number in the upload
+       // record instead of the model name, to match the USB version
+       sessionInfo.deviceModel = cfg.deviceInfo.model;
+     }
 
      cfg.api.upload.toPlatform(
        data.post_records, sessionInfo, progress, cfg.groupId,
