+++ lib/drivers/roche/accuChekUSB.js
@@ -35,6 +35,8 @@ import {
   getProductionSpecEntry,
 } from './utils';
 
+import getModelName from './models';
+
 const isBrowser = typeof window !== 'undefined';
 // eslint-disable-next-line no-console
 const debug = isBrowser ? require('bows')('AccuChekUSBDriver') : console.log;
@@ -506,27 +508,6 @@ module.exports = (config) => {
   // With no date & time settings changes available,
   // timezone is applied across-the-board
   cfg.tzoUtil = new TZOUtil(cfg.timezone, new Date().toISOString(), []);
-
-  const models = [
-    { name: 'Aviva Connect', numbers: [483, 484, 497, 498, 499, 500, 502, 685] },
-    { name: 'Performa Connect', numbers: [479, 501, 503, 765] },
-    { name: 'Guide', numbers: [921, 922, 923, 925, 926, 929, 930, 932] },
-    { name: 'Instant (single-button)', numbers: [958, 959, 960, 961, 963, 964, 965] },
-    { name: 'Guide Me', numbers: [897, 898, 901, 902, 903, 904, 905] },
-    { name: 'Instant (two-button)', numbers: [972, 973, 975, 976, 977, 978, 979, 980] },
-    { name: 'Instant S (single-button)', numbers: [966, 967, 968, 969, 970, 971] },
-    { name: 'ReliOn Platinum', numbers: [982] },
-  ];
-
-  const getModelName = (number) => {
-    // eslint-disable-next-line no-restricted-syntax
-    for (const i in models) {
-      if (models[i].numbers.indexOf(_.toInteger(number)) >= 0) {
-        return models[i].name;
-      }
-    }
-    return `Unknown model ${number}`;
-  };
 
   return {
     /* eslint no-param-reassign:
